refactor(site): SSR cron revalidate route + all-season prerender (G)
Migrate the cache-invalidation endpoint off the Pages Router onto an App
Router route handler, completing the pure-App-Router topology (src/pages is
now gone).

- app/api/cron/revalidate/route.ts: POST handler, same /api/cron/revalidate
  path so REVALIDATE_URL is unchanged. Swap updateTag -> revalidateTag(tag,
  'max'). The ingest changes data out-of-band, so background
  stale-while-revalidate (not read-your-writes) is correct; 'max' matches the
  cacheLife profile the data fetchers use. Fail-closed 500 when CRON_SECRET
  is unset, 401 on mismatch.
- app/api/cron/ingest/route.ts: re-export POST as the legacy alias the old
  pages route provided.
- [season]/page.tsx: generateStaticParams now enumerates every season
  (getSeasons) instead of only the current year, so all seasons prerender.

Verified (dev, live API): no-auth 401, wrong-token 401, correct-token 200
with all six tags revalidated on both /revalidate and /ingest; full build
174/174 static pages (was 97), /2015 and /1995 render standings+schedule.
